+++ engine/src/entities/music/BPM.cpp
@@ -6,8 +6,8 @@ namespace XJ {
 
   static double NANOS_PER_SECOND = 1000000000;
 
-  long BPM::beatsNanos(long beats, float bpm) {
-    return (long) (NANOS_PER_SECOND * beats * 60 / bpm);
+  long long BPM::beatsNanos(float beats, float bpm) {
+    return (long long) (NANOS_PER_SECOND * beats * 60 / bpm);
   }
 
   float BPM::velocity(float bpm) {
